requirements.txt
livekit-agents==1.4.1
livekit-plugins-openai
livekit-plugins-silero
python-dotenv
pinecone>=3.0.0
langchain>=0.2.0
langchain-openai
langchain-pinecone
tiktoken
